Stage

+++ src/TD.Office/TD.Office.Common/TD.Office.Common.Contracts/Entities/UslovFormiranjaWebCeneEntity.cs
@@ -5,7 +5,7 @@ namespace TD.Office.Common.Contracts.Entities
 {
     public class UslovFormiranjaWebCeneEntity : LSCoreEntity
     {
-        public int WebProductId { get; set; }
+        public long WebProductId { get; set; }
         public UslovFormiranjaWebCeneType Type { get; set; }
         public decimal Modifikator { get; set; }
     }
